pyproject.toml
[project]
name = "if2150-2025-k02-g11-foodshare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "flask>=3.1.2",
    "mysql-connector-python>=9.5.0",
    "pillow>=12.0.0",
    "requests>=2.32.5",
]
